Version 2.6.3

+++ system/modules/newsletter/config/database.sql
@@ -26,7 +26,7 @@ CREATE TABLE `tl_newsletter` (
   `template` varchar(32) NOT NULL default '',
   `sendText` char(1) NOT NULL default '',
   `sent` char(1) NOT NULL default '',
-  `date` int(10) unsigned NOT NULL default '0',
+  `date` int(11) NOT NULL default '0',
   PRIMARY KEY  (`id`),
   KEY `pid` (`pid`)
 ) ENGINE=MyISAM DEFAULT CHARSET=utf8;
@@ -42,7 +42,7 @@ CREATE TABLE `tl_newsletter_channel` (
   `id` int(10) unsigned NOT NULL auto_increment,
   `tstamp` int(10) unsigned NOT NULL default '0',
   `title` varchar(255) NOT NULL default '',
-  `jumpTo` smallint(5) unsigned NOT NULL default '0'
+  `jumpTo` int(10) unsigned NOT NULL default '0'
   PRIMARY KEY  (`id`)
 ) ENGINE=MyISAM DEFAULT CHARSET=utf8;
